move actions to publication reducer
