+++ apps/chat/src/store/prompts/prompts.reducers.ts
@@ -1,6 +1,6 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
 
-import { combineEntities, sortAllVersions } from '@/src/utils/app/common';
+import { combineEntities } from '@/src/utils/app/common';
 import { constructPath } from '@/src/utils/app/file';
 import {
   addGeneratedFolderId,
@@ -19,7 +19,6 @@ import { translate } from '@/src/utils/app/translation';
 import { FeatureType, UploadStatus } from '@/src/types/common';
 import { FolderInterface, FolderType } from '@/src/types/folder';
 import { Prompt, PromptInfo } from '@/src/types/prompt';
-import { PublicVersionGroups } from '@/src/types/publication';
 import { SearchFilters } from '@/src/types/search';
 import '@/src/types/share';
 
@@ -28,8 +27,6 @@ import { DEFAULT_FOLDER_NAME } from '@/src/constants/default-ui-settings';
 import * as PromptsSelectors from './prompts.selectors';
 import { PromptsState } from './prompts.types';
 
-import omit from 'lodash-es/omit';
-import uniqBy from 'lodash-es/uniqBy';
 import xor from 'lodash-es/xor';
 
 export { PromptsSelectors };
@@ -51,7 +48,6 @@ const initialState: PromptsState = {
   isNewPromptCreating: false,
   chosenPromptIds: [],
   chosenEmptyFoldersIds: [],
-  publicVersionGroups: {},
 };
 
 export const promptsSlice = createSlice({
@@ -470,93 +466,6 @@ export const promptsSlice = createSlice({
       state.chosenEmptyFoldersIds = xor(
         state.chosenEmptyFoldersIds,
         payload.ids,
-      );
-    },
-    addPublicVersionGroups: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        publicVersionGroups: PublicVersionGroups;
-      }>,
-    ) => {
-      for (const key in payload.publicVersionGroups) {
-        const selectedVersion =
-          payload.publicVersionGroups[key]?.selectedVersion ||
-          state.publicVersionGroups[key]?.selectedVersion;
-
-        if (selectedVersion) {
-          state.publicVersionGroups[key] = {
-            selectedVersion,
-            allVersions: sortAllVersions(
-              uniqBy(
-                [
-                  ...(state.publicVersionGroups[key]?.allVersions || []),
-                  ...(payload.publicVersionGroups[key]?.allVersions || []),
-                ],
-                'id',
-              ),
-            ),
-          };
-        }
-      }
-    },
-    setNewVersionForPublicVersionGroup: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        versionGroupId: string;
-        newVersion: NonNullable<PublicVersionGroups[string]>['selectedVersion'];
-        oldVersion: NonNullable<PublicVersionGroups[string]>['selectedVersion'];
-      }>,
-    ) => {
-      // link to state.publicVersionGroups[payload.versionGroupId]
-      const versionGroup = state.publicVersionGroups[payload.versionGroupId];
-
-      if (versionGroup) {
-        versionGroup.selectedVersion = payload.newVersion;
-      }
-    },
-    removePublicVersionGroups: (
-      state,
-      {
-        payload,
-      }: PayloadAction<{
-        groupsToRemove: {
-          versionGroupId: string;
-          groupIds: string[];
-        }[];
-      }>,
-    ) => {
-      // versionGroups it's a link to state.publicVersionGroups[payload.versionGroupId]
-      const groupWithIdsToRemove = payload.groupsToRemove.map((group) => ({
-        versionGroup: state.publicVersionGroups[group.versionGroupId],
-        idsToRemove: group.groupIds,
-        versionGroupId: group.versionGroupId,
-      }));
-
-      groupWithIdsToRemove.forEach(
-        ({ versionGroup, idsToRemove, versionGroupId }) => {
-          if (versionGroup) {
-            const filteredVersionGroups = versionGroup.allVersions.filter(
-              (group) => !idsToRemove.includes(group.id),
-            );
-
-            versionGroup.allVersions = filteredVersionGroups;
-
-            if (idsToRemove.includes(versionGroup.selectedVersion.id)) {
-              if (filteredVersionGroups[0]) {
-                versionGroup.selectedVersion = filteredVersionGroups[0];
-              } else {
-                state.publicVersionGroups = omit(
-                  state.publicVersionGroups,
-                  versionGroupId,
-                );
-              }
-            }
-          }
-        },
       );
     },
   },
