+++ apps/chat/src/store/prompts/prompts.types.ts
@@ -1,6 +1,5 @@
 import { FolderInterface } from '@/src/types/folder';
 import { PromptInfo } from '@/src/types/prompt';
-import { PublicVersionGroups } from '@/src/types/publication';
 import { SearchFilters } from '@/src/types/search';
 
 export interface PromptsState {
@@ -20,5 +19,4 @@ export interface PromptsState {
   isNewPromptCreating: boolean;
   chosenPromptIds: string[];
   chosenEmptyFoldersIds: string[];
-  publicVersionGroups: PublicVersionGroups;
 }
